requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
chromadb>=0.5.5
sentence-transformers>=3.0.1
transformers>=4.44.0
torch>=2.2.0
Pillow>=10.4.0
networkx>=3.3
numpy>=1.26.4
pandas>=2.2.2
scikit-learn>=1.5.1
tqdm>=4.66.4
openai>=1.40.0
pytest>=8.3.2
